New license
Users get the first 7 days to use the product free of giving up ANY information. After these 7 days, users can optionally create an account (just email) to get an extra 23 days free. This makes a scaffold for users to ONLY have to make a payment to regain access to the product after the 30 days if they make it that far